validation_utils/requirements.txt
requests==2.21.0
celery==4.4.2
sentry_sdk==1.5.12
psycopg2==2.9.3
python-dotenv==0.20.0